requirements.txt
Django>=2.0,<3.0
psycopg2-binary>=2.8
djangorestframework>=3.12
django-cors-headers>=2.4.0
#django-cors-middleware
#django-cors-headers==1.1.0

#Django>=1.8,<1.9
#djangorestframework==3.0.5

# dev deps
factory-boy
